reorganize fct tests
