+++ testingDataset/expected_info_files/expected_hashes.json
@@ -1,4 +1,5 @@
 {
     "mybasicpangenome/gene_families.tsv": "9d6219523e890b08c467c936590b37436e915840a797ea161c9707041c3b821c",
-    "stepbystep/gene_families.tsv": "9d6219523e890b08c467c936590b37436e915840a797ea161c9707041c3b821c"
+    "stepbystep/gene_families.tsv": "9d6219523e890b08c467c936590b37436e915840a797ea161c9707041c3b821c",
+    "myannopang/gene_families.tsv": "5cb88afbc1a49e7d21531d36d5ee3471a87834f4bb82578347dc92231ba0a452"
 }
