+++ tests/pytest.ini
@@ -2,3 +2,6 @@
 markers =
 	nucleotides: marks tests to run with nucleotides alphabets
 	aminoacids: marks tests to run with amino-acids alphabets
+
+	log_cli = true
+	log_level = INFO
